+++ app/src/main/java/com/example/bookingagent/data/db/entities/Service.kt
@@ -13,5 +13,5 @@ data class Service(
 	val id: Int,
 	val name: String,
 	val description: String,
-	val price: Int
+	val price: Float
 ) : Parcelable
